bot/requirements.txt
discord.py>=2.3.0
firebase-admin>=6.5.0
python-dotenv>=1.0.0
